requirements.txt
Flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.9
bcrypt==4.2.1
PyJWT==2.10.1
gunicorn==23.0.0
python-dotenv==1.0.1